move team data to main drill_data file
