+++ app/content/team/users/[user]/drills/[id].js
@@ -1,9 +1,6 @@
 //assumed to be statistics
 import React from "react";
 import { useLocalSearchParams } from "expo-router";
-
- // At some point migrate this to team_data.json after adding enough fake data
- // Also, maybe move pass drill data as prop
 import drillData from "~/drill_data.json";
 import BarChartScreen from "~/components/barChart";
 
@@ -14,7 +11,7 @@ export default function Stat() {
   return (
     <>
       <BarChartScreen
-        drillData={drillData["users"]["1"]["history"][drill_id]}
+        drillData={drillData["users"][user_id]["history"][drill_id]}
         mainOutputAttempt={drillData["drills"][drill_id]["mainOutputAttempt"]}
       />
     </>
