+++ app/content/team/users/[user]/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import drillData from "~/team_data.json";
+import drillData from "~/drill_data.json";
 import ProfileCard from "~/components/profileCard";
 import { ScrollView, Text } from "react-native";
 import DrillCard from "~/components/drillCard";
